Remove need to explicitly fine-tune resolution for individual plugins
By applying the jvm-ecosystem plugin and specifying the request attributes to be the same as the attributes that Gradle
uses during normal plugin resolution, we avoid needing to fine-tune resolution for individual plugins.

As such, we can remove all usages of the 'configuration' DSL from the :runner project. We have also removed the
'shadowed' method from the DSL, as it is no longer requried to be specified explicitly. In general, we should be
able to remove the 'configuration' DSL block entirely, as it is no longer called by any of the currently analyzed
plugins, and will likely not be needed for any other conventional plugin that is already resolvable by Gradle.